6 May 2020, 2:00 pm
Altered the (0) repeat data set to remove columns 4 and 5 that had no data, but the .csv was loading as if it had 5 columns of data.

Cleaned up beetle_script code to remove duplicate lines of code, unneccesary comments, and cleaned up the organization. Put alternative fig 5 code hashed out in its own section in case we return to it later. Will decide whether to remove from this version once other authors have signed off on the figures.